fix: remove authentication components & adjust gradle settings
As suggested in the lab manual, use of jwt and auth in gradle must be removed given that we decided not to pass Jwt tokens.
Removing these from the build scripts necessitates removal of sample auth components, since project does not compile otherwise.

Fixes the issue where any request to an endpoint without a proper Jwt token resulted in a 401 Unauthorized